updates to api
